requirements.txt
amqp==5.3.1
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
black==26.1.0
boto3==1.42.54
botocore==1.42.54
cbor2==5.8.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coverage==7.13.4
cron_descriptor==2.0.6
cryptography==46.0.5
daphne==4.2.1
distlib==0.4.0
Django==5.2.11
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-filter==25.2
django-redis==6.0.0
django-storages==1.14.6
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
factory_boy==3.3.3
Faker==40.4.0
filelock==3.24.3
flake8==7.3.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
identify==2.6.16
idna==3.11
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
isort==8.0.0
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
librt==0.8.1
mccabe==0.7.0
msgpack==1.1.2
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
pathspec==1.0.4
paystack-python==0.1
paystack-sdk==1.0.1
pillow==12.1.1
platformdirs==4.9.2
pluggy==1.6.0
pre_commit==4.5.1
prompt_toolkit==3.0.52
psycopg==3.3.3
psycopg-binary==3.3.3
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-django==4.12.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytokens==0.4.1
PyYAML==6.0.3
redis==7.2.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
sentry-sdk==2.53.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
vine==5.1.0
virtualenv==20.38.0
wcwidth==0.6.0
whitenoise==6.11.0
zope.interface==8.2
